print file hash and move connections into connection
