+++ cmb_protocol/constants.py
@@ -2,3 +2,4 @@ MAXIMUM_TRANSMISSION_UNIT = 512
 SYMBOLS_PER_BLOCK = 100
 DEFAULT_PORT = 9999
 DEFAULT_IP_ADDR = '127.0.0.1'
+RESOURCE_ID_STRUCT_FORMAT = '!16sQ'
